Move MBROLA scripts to root folder
